requirements.txt
pandas
numpy
requests
openpyxl
yfinance
scikit-learn
torch
google-generativeai
python-dotenv